check added

+++ backend/api/controller/task.js
@@ -1,9 +1,12 @@
 const taskSchema =require('../../models/Task')
+const userSchema=require('../../models/User')
 exports.addTask=async(req,res,next)=>{
     try {
         const userId=req.userData.userId;
         const {name,isDone}=req.body;
         if(!name) return res.status(400).json({status:false,message:"Invalid Data"});
+        const isValidUser=await userSchema.findOne({_id:userId})
+        if(!isValidUser) return res.status(400).json({status:false,message:"User Doesn't Exists"})
         const newTask=new taskSchema({
             userId:userId,
             name:name,
@@ -26,6 +29,8 @@ exports.addTask=async(req,res,next)=>{
 exports.getalltask=async(req,res,next)=>{
     try {
         const userId=req.userData.userId;
+        const isValidUser=await userSchema.findOne({_id:userId})
+        if(!isValidUser) return res.status(400).json({status:false,message:"User Doesn't Exists"})
         const tasks=await taskSchema.find({userId:userId}).select("_id name isDone")
         tasks.sort((taskA, taskB) => {
             if (taskA.isDone === false && taskB.isDone === true) {
@@ -52,6 +57,8 @@ exports.updatetask=async(req,res,next)=>{
       try {
         const {name,isDone,_id}=req.body;
         if(!_id) return res.status(400).json({status:false,message:"Invalid Data"});
+        const isValidUser=await userSchema.findOne({_id:userId})
+        if(!isValidUser) return res.status(400).json({status:false,message:"User Doesn't Exists"})
         let updatedFiled={}
         if(name) updatedFiled.name=name;
         if(isDone ==true || isDone==false) updatedFiled.isDone=isDone;
@@ -70,6 +77,8 @@ exports.deletetask=async(req,res,next)=>{
     try {
         const {_id}=req.body;
         if(!_id) return res.status(400).json({status:false,message:"Invalid Data"});
+        const isValidUser=await userSchema.findOne({_id:userId})
+        if(!isValidUser) return res.status(400).json({status:false,message:"User Doesn't Exists"})
         const deletedTask=await taskSchema.findOneAndDelete({_id:_id},{new:true})
         if(!deletedTask) return res.status(400).json({status:false,message:"Task not found"})
         return res.status(204).json({status:true,message:"Task Successfully Deleted"})
